requirements.txt
adal==1.0.1
ajaxuploader==0.3.8
altgraph==0.15
amqp==2.3.2
anyjson==0.3.3
appnope==0.1.0
arrow==0.12.1
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
awscli==1.15.40
awsebcli==3.14.1
azure==3.0.0
azure-batch==4.1.3
azure-common==1.1.12
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.3
azure-datalake-store==0.0.22
azure-eventgrid==0.1.0
azure-graphrbac==0.40.0
azure-keyvault==0.3.7
azure-mgmt==2.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==0.2.0
azure-mgmt-billing==0.1.0
azure-mgmt-cdn==2.0.0
azure-mgmt-cognitiveservices==2.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==3.0.1
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==0.3.1
azure-mgmt-containerregistry==1.0.1
azure-mgmt-containerservice==3.0.1
azure-mgmt-cosmosdb==0.3.1
azure-mgmt-datafactory==0.4.0
azure-mgmt-datalake-analytics==0.3.0
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.3.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==1.2.0
azure-mgmt-documentdb==0.1.3
azure-mgmt-eventgrid==0.4.0
azure-mgmt-eventhub==1.2.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iothub==0.4.0
azure-mgmt-iothubprovisioningservices==0.1.0
azure-mgmt-keyvault==0.40.0
azure-mgmt-loganalytics==0.1.0
azure-mgmt-logic==2.1.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementpartner==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==0.2.0
azure-mgmt-monitor==0.4.0
azure-mgmt-msi==0.1.0
azure-mgmt-network==1.7.1
azure-mgmt-notificationhubs==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-powerbiembedded==1.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-recoveryservices==0.2.0
azure-mgmt-recoveryservicesbackup==0.1.1
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.1.0
azure-mgmt-resource==1.2.2
azure-mgmt-scheduler==1.1.3
azure-mgmt-search==1.0.0
azure-mgmt-servermanager==1.2.0
azure-mgmt-servicebus==0.4.0
azure-mgmt-servicefabric==0.1.0
azure-mgmt-sql==0.8.6
azure-mgmt-storage==1.5.0
azure-mgmt-subscription==0.1.0
azure-mgmt-trafficmanager==0.40.0
azure-mgmt-web==0.34.1
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==6.1.2.9
azure-servicemanagement-legacy==0.20.6
azure-storage==0.36.0
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-file==1.1.0
azure-storage-nspkg==3.0.0
azure-storage-queue==1.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
bdist-mpkg==0.5.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
bitcoin==1.1.42
bleach==2.1.3
blessed==1.14.2
boto==2.48.0
botocore==1.10.40
cached-property==1.4.3
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
channels==1.1.5
characteristic==14.3.0
chardet==3.0.4
click==6.7
cobble==0.1.3
CodernityDB==0.5.0
colorama==0.3.9
configobj==4.7.2
configparser==3.5.0
confusable-homoglyphs==3.1.1
constantly==15.1.0
coverage==4.5.1
crispy-forms-foundation==0.6.4
cryptography==2.2.2
csscompressor==0.9.5
cssselect==1.0.3
cssutils==1.0.2
Cython==0.28.3
daphne==1.4.2
decorator==4.3.0
demjson==2.2.4
dj-database-url==0.5.0
dj-upload-to==1.1
Django==1.11.13
django-ajax-search==1.5.1
django-ajaximage==0.2.8
django-allauth==0.36.0
django-appconf==1.0.1
django-apptemplates==1.4
django-async==0.7.3
django-autocomplete-light==3.2.10
django-block-ip==0.2.0
django-classy-tags==0.8.0
django-compressor==2.2
django-contrib-comments==1.8.0
django-crispy-forms==1.7.2
django-el-pagination==3.2.4
django-facebook==6.0.3
django-htmlmin==0.10.0
django-image-cropper==0.3.0
django-image-cropping==1.1.0
django-ipware==2.1.0
django-jcrop==0.0.2
django-js-asset==1.1.0
django-load==1.0.0
django-model-utils==3.1.2
django-mptt==0.9.0
django-multiselectfield==0.1.8
django-mysql==2.2.2
django-notifications-hq==1.4.0
django-nyt==1.0
django-picklefield==1.0.0
django-pipeline==1.6.14
django-polymorphic==2.0.2
django-q==0.9.4
django-recaptcha==1.4.0
django-redis-sessions==0.6.1
django-registration==2.4.1
django-registration-redux==2.4
django-render-partial==0.3
django-reversion==2.0.13
django-sekizai==0.10.0
django-ses==0.8.5
django-smtp-ssl==1.0
django-social-auth==0.7.28
django-solo==1.1.3
django-static-sitemaps==4.4.0
django-statici18n==1.8.2
django-storages==1.6.6
django-threadedcomments==1.2
django-tinymce==2.7.0
django-websocket-redis==0.5.2
dnspython==1.15.0
docker==3.3.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dodge==0.1.9
easy-thumbnails==2.5
EasyProcess==0.2.3
ecdsa==0.13
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gevent==1.3.3
gevent-socketio==0.3.6
gevent-websocket==0.10.1
gipc==0.6.0
Glances==2.11.1
gnureadline==6.3.8
greenlet==0.4.13
gunicorn==19.8.1
haikunator==2.1.0
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.11.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
iniparse==0.4
interruptingcow==0.8
ipcalc==1.99.0
ipfsapi==0.4.3
IPy==0.75
ipython==5.7.0
ipython-genutils==0.2.0
isodate==0.6.0
jmespath==0.9.3
jsmin==2.2.2
json-rpc==1.11.0
jsonfield==2.0.2
jsonpickle==0.9.6
jsonschema==2.6.0
keyring==13.0.0
keyrings.alt==3.1
kombu==4.2.1
leveldb==0.194
lmdb==0.94
lockfile==0.12.2
lxml==4.2.1
macholib==1.9
Markdown==2.6.11
modulegraph==0.16
msgpack-python==0.5.6
msrest==0.4.5
msrestazure==0.4.32
mysql-connector==2.1.6
MySQL-python==1.2.5
mysqlclient==1.3.12
ndg-httpsclient==0.5.0
newrelic==3.2.2.94
notifications==0.3.2
oauth2==1.9.0.post1
oauthlib==2.1.0
olefile==0.45.1
paramiko==2.4.1
parsel==1.4.0
parsimonious==0.8.0
pathlib==1.0.1
pathlib2==2.3.2
pathspec==0.5.5
pbkdf2==1.3
perf==0.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.6
psycogreen==1.0
ptyprocess==0.5.2
py==1.5.3
py-ecc==1.4.2
py2-ipaddress==3.4.1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.1
pycryptodomex==3.6.1
pycurl==7.19.0
PyDispatcher==2.0.5
pyelliptic==1.5.8
Pygments==2.2.0
pygpgme==0.3
pyjwkest==1.4.0
PyJWT==1.6.4
pyliblzma==0.5.3
pylru==1.1.0
PyNaCl==1.2.1
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pysha3==1.0.2
pysqlite==2.8.3
python-augeas==0.5.0
python-dateutil==2.7.3
python-magic==0.4.15
python-openid==2.2.5
pytz==2018.4
pyudev==0.21.0
PyVirtualDisplay==0.2.1
pyxattr==0.5.1
pyxdg==0.26
PyYAML==3.12
queuelib==1.5.0
rcssmin==1.0.6
redis==2.10.6
repoze.lru==0.7
requests==2.18.4
requests-oauthlib==1.0.0
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.13
scandir==1.7
SecretStorage==2.3.1
selenium==3.12.0
semantic-version==2.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.15.0
singledispatch==3.4.0.3
six==1.11.0
sorl-thumbnail==12.4.1
South==1.0.2
sphinx-me==0.3
sshtunnel==0.1.4
statistics==1.0.3.5
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
timeout-decorator==0.4.0
tinyrpc==0.9.1
tld==0.9
tornado==5.0.2
tornado-redis==2.4.18
tox==3.0.0
traitlets==4.3.2
Twisted==18.4.0
txaio==2.10.0
typing==3.6.4
Unidecode==1.0.22
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
virtualenv==16.0.0
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==0.14.1
wsaccel==0.6.2
xattr==0.9.3
xvfbwrapper==0.2.9
zope.interface==4.5.0
